requirements.txt
datasets==1.12.1
pyannote.audio==2.2
fastapi==0.68.0
pydub==0.25.1
numpy==1.21.0
TTS==0.0.3
transformers==4.12.2